Factor repr_write into repr
